# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoveZone API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  id: number;
  username: string;
}

export interface DashboardStats {
  totalListings: number;
  totalStates: number;
  totalCities: number;
  totalSeoPages: number;
  featuredListings: number;
  activeListings: number;
}

export interface SiteSettings {
  id: number;
  siteName: string;
  phoneNumber: string;
  whatsappNumber: string;
  email: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface SiteSettingsUpdate {
  siteName?: string;
  phoneNumber?: string;
  whatsappNumber?: string;
  email?: string;
}

export interface State {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  metaTitle?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  listingCount: number;
  /** @nullable */
  createdAt?: string | null;
}

export interface StateInput {
  name: string;
  slug: string;
  description?: string;
  metaTitle?: string;
  metaDescription?: string;
}

export interface StateUpdate {
  name?: string;
  slug?: string;
  description?: string;
  metaTitle?: string;
  metaDescription?: string;
}

export interface City {
  id: number;
  name: string;
  slug: string;
  stateSlug: string;
  stateName: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  metaTitle?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  listingCount: number;
  /** @nullable */
  createdAt?: string | null;
}

export interface CityInput {
  name: string;
  slug: string;
  stateSlug: string;
  description?: string;
  metaTitle?: string;
  metaDescription?: string;
}

export interface CityUpdate {
  name?: string;
  slug?: string;
  stateSlug?: string;
  description?: string;
  metaTitle?: string;
  metaDescription?: string;
}

export interface Listing {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  stateSlug: string;
  citySlug: string;
  stateName: string;
  cityName: string;
  /** @nullable */
  services?: string | null;
  /** @nullable */
  height?: string | null;
  /** @nullable */
  weight?: string | null;
  isActive: boolean;
  isFeatured: boolean;
  /** @nullable */
  createdAt?: string | null;
}

export interface ListingInput {
  name: string;
  slug: string;
  age?: number;
  description?: string;
  phone?: string;
  whatsapp?: string;
  imageUrl?: string;
  stateSlug: string;
  citySlug: string;
  services?: string;
  height?: string;
  weight?: string;
  isActive?: boolean;
  isFeatured?: boolean;
}

export interface ListingUpdate {
  name?: string;
  slug?: string;
  age?: number;
  description?: string;
  phone?: string;
  whatsapp?: string;
  imageUrl?: string;
  stateSlug?: string;
  citySlug?: string;
  services?: string;
  height?: string;
  weight?: string;
  isActive?: boolean;
  isFeatured?: boolean;
}

export interface ListingPage {
  data: Listing[];
  total: number;
  page: number;
  limit: number;
}

export interface SeoPage {
  id: number;
  keyword: string;
  slug: string;
  seoTitle: string;
  metaDescription: string;
  h1Heading: string;
  content: string;
  /** @nullable */
  faq?: string | null;
  isActive?: boolean;
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}

export interface SeoPageInput {
  keyword: string;
  slug: string;
  seoTitle: string;
  metaDescription: string;
  h1Heading: string;
  content: string;
  faq?: string;
  isActive?: boolean;
}

export interface SeoPageUpdate {
  keyword?: string;
  slug?: string;
  seoTitle?: string;
  metaDescription?: string;
  h1Heading?: string;
  content?: string;
  faq?: string;
  isActive?: boolean;
}

export type ListCitiesParams = {
stateSlug?: string;
};

export type ListListingsParams = {
stateSlug?: string;
citySlug?: string;
featured?: boolean;
page?: number;
limit?: number;
};