# lib/api-zod/src/generated/types/city.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoveZone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface City {
  id: number;
  name: string;
  slug: string;
  stateSlug: string;
  stateName: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  metaTitle?: string | null;
  /** @nullable */
  metaDescription?: string | null;
  listingCount: number;
  /** @nullable */
  createdAt?: string | null;
}