# lib/api-zod/src/generated/types/cityInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoveZone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CityInput {
  name: string;
  slug: string;
  stateSlug: string;
  description?: string;
  metaTitle?: string;
  metaDescription?: string;
}