# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoveZone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalListings: number;
  totalStates: number;
  totalCities: number;
  totalSeoPages: number;
  featuredListings: number;
  activeListings: number;
}