# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoveZone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Listing {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  stateSlug: string;
  citySlug: string;
  stateName: string;
  cityName: string;
  /** @nullable */
  services?: string | null;
  /** @nullable */
  height?: string | null;
  /** @nullable */
  weight?: string | null;
  isActive: boolean;
  isFeatured: boolean;
  /** @nullable */
  createdAt?: string | null;
}