# lib/api-zod/src/generated/types/listingPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoveZone API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Listing } from './listing';

export interface ListingPage {
  data: Listing[];
  total: number;
  page: number;
  limit: number;
}