# lib/api-zod/src/generated/types/listingUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoveZone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ListingUpdate {
  name?: string;
  slug?: string;
  age?: number;
  description?: string;
  phone?: string;
  whatsapp?: string;
  imageUrl?: string;
  stateSlug?: string;
  citySlug?: string;
  services?: string;
  height?: string;
  weight?: string;
  isActive?: boolean;
  isFeatured?: boolean;
}