# lib/api-zod/src/generated/types/seoPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoveZone API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SeoPage {
  id: number;
  keyword: string;
  slug: string;
  seoTitle: string;
  metaDescription: string;
  h1Heading: string;
  content: string;
  /** @nullable */
  faq?: string | null;
  isActive?: boolean;
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}